+++ src/Lidarr.Plugin.Slskd/Models/TransferDirection.cs
@@ -1,0 +1,7 @@
+﻿namespace NzbDrone.Plugin.Slskd.Models;
+
+public enum TransferDirection
+{
+    Download,
+    Upload
+}
