+++ src/Lidarr.Plugin.Slskd/Models/TransferStates.cs
@@ -1,0 +1,7 @@
+﻿namespace NzbDrone.Plugin.Slskd.Models;
+
+public class TransferStates
+{
+    public TransferStateEnum State { get; set; }
+    public TransferStateEnum Substate { get; set; }
+}
